+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 > - :house: [Internal]
 
 ## [Unreleased]
+### :rocket: New Feature
+- `common`
+  - color-picker component initial implementation
 <hr/>
 
 ## 3.0.8  (Mar 12, 2019)